use table actions component in transactions list
